splits round_robin integration testing into separate file
